src/requirements.txt
numpy
matplotlib
pandas
py-sudoku